introduced sharedworkspaceexception, made producer creation concurrent, established notification of workspace changes

+++ src/at/ac/tuwien/complang/sbc11/mozart/StandaloneServer.java
@@ -4,7 +4,7 @@ import org.mozartspaces.core.Server;
 
 public class StandaloneServer {
 
-	public static final long SERVER_PORT = 1234;
+	public static final int SERVER_PORT = 1234;
 	/**
 	 * @param args
 	 */
